requirements.txt
fastapi
uvicorn
torch==2.1.0+cpu
torchvision==0.16.0+cpu
pillow
python-multipart
numpy<2
--extra-index-url https://download.pytorch.org/whl/cpu